Drop sword fix, Added a better way to play sounds, Removed duplicate classes in one object, Most of the things are working now

+++ GMDRPGGame/Assets/Scripts/Control/PlayerController.cs
@@ -92,28 +92,40 @@ namespace RPG.Control
             }
             else if (Input.GetKeyDown(KeyCode.Q))
             {
-                //health.Heal(50);
-                useItem.CheckWhatItemToUse(inventory.GetItemList()[0].itemType);
-                inventory.DeleteItem(0);
-                uiInventory.RefreshInventoryItems();
+                if (inventory.GetItemList().Count > 0)
+                {
+                    useItem.CheckWhatItemToUse(inventory.GetItemList()[0].itemType);
+                    inventory.DeleteItem(0);
+                    uiInventory.RefreshInventoryItems();
+                }
+
             }
             else if (Input.GetKeyDown(KeyCode.W))
             {
-                useItem.CheckWhatItemToUse(inventory.GetItemList()[1].itemType);
-                inventory.DeleteItem(1);
-                uiInventory.RefreshInventoryItems();
+                if (inventory.GetItemList().Count > 1)
+                {
+                    useItem.CheckWhatItemToUse(inventory.GetItemList()[1].itemType);
+                    inventory.DeleteItem(1);
+                    uiInventory.RefreshInventoryItems();
+                }
             }
             else if (Input.GetKeyDown(KeyCode.E))
             {
-                useItem.CheckWhatItemToUse(inventory.GetItemList()[2].itemType);
-                inventory.DeleteItem(2);
-                uiInventory.RefreshInventoryItems();
+                if (inventory.GetItemList().Count > 2)
+                {
+                    useItem.CheckWhatItemToUse(inventory.GetItemList()[2].itemType);
+                    inventory.DeleteItem(2);
+                    uiInventory.RefreshInventoryItems();
+                }
             }
             else if (Input.GetKeyDown(KeyCode.R))
             {
-                useItem.CheckWhatItemToUse(inventory.GetItemList()[3].itemType);
-                inventory.DeleteItem(3);
-                uiInventory.RefreshInventoryItems();
+                if (inventory.GetItemList().Count > 3)
+                {
+                    useItem.CheckWhatItemToUse(inventory.GetItemList()[3].itemType);
+                    inventory.DeleteItem(3);
+                    uiInventory.RefreshInventoryItems();
+                }
             }
 
             updateHUD();
